Digital_Ocean_backend/normalization/requirements.txt
asyncpg>=0.29.0
python-dotenv
